Fix kotest5 test
This commit fixes the kotest 5 test: it should have lived in the
functionalTests project, because it makes use of an external, optional
plugin: those tests are intentionally isolated, so that we can verify
that the code we write works if classes are missing.
